requirements.txt
selenium==4.35.0
beautifulsoup4==4.12.3
lxml==5.3.0
langchain==0.3.27
openai==1.109.1
python-dotenv==1.1.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-api-python-client==2.183.0
